requirements.txt
transformers==4.38.2
peft==0.8.2
accelerate==0.28.0
datasets==2.18.0
bitsandbytes==0.42.0
sentencepiece
safetensors
huggingface_hub
trl==0.7.4
torch==2.1.0
fastapi
uvicorn[standard]==0.29.0
pydantic
python-docx
azure-identity
azure-ai-ml >= 1.13.0
azure-cosmos
azure-storage-blob
python-dotenv
python-multipart
tqdm
beautifulsoup4
protobuf
requests
numpy<2.0.0
openai